azdocs-gpt-backend/requirements.txt
# requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
flask-cors==3.0.10
openai==1.3.0
azure-identity==1.23.0
azure-search-documents==11.5.2
azure-core==1.29.0
httpx==0.27.0
